\name{EnsDbLite.Hsapiens.84}
\docType{package}

\alias{EnsDbLite.Hsapiens.84-package}
\alias{EnsDbLite.Hsapiens.84}
\alias{EnsDbLite.Hsapiens.84}


\title{Ensembl-based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Tue Apr 12 13:53:44 2016 and based on the GRCh38
}

\author{.}



\examples{
## load the library
##library(EnsDbLite.Hsapiens.84)
## list the contents that are loaded into memory
ls('package:EnsDbLite.Hsapiens.84')
## show the db object that is loaded by calling it's name
EnsDbLite.Hsapiens.84

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
